reorganize tests into groups
